pyproject.toml
[tool.poetry]
name = "etl"
version = "0.1.0"
description = ""
authors = ["Nikita <zigman.nikita@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
pydantic = "^1.10.7"
psycopg2-binary = "^2.9.6"
loguru = "^0.7.0"
python-dotenv = "^1.0.0"
urllib3 = "1.26.15"
elasticsearch = "^8.7.0"
redis = "^4.5.5"
backoff = "^2.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
ipython = "^8.13.2"
mypy = "^1.3.0"
pre-commit = "^3.3.2"
flake8 = "^6.0.0"
isort = "^5.12.0"
black = "^23.3.0"
types-redis = "^4.5.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 79
exclude = '''
/(
    \.git
  | \.mypy_cache
  | venv
  | \.venv
)/
'''

[tool.mypy]
files = "src/"
ignore_missing_imports = "True"   
strict_optional = "True"

[tool.bandit]
exclude_dirs = ["*/tests"]

[tool.isort]
profile = "black"
line_length = 79

[tool.flake8]
ignore = ['E501']
exclude = [
    ".git",
    "__pycache__",
    "venv",
    "migrations"
]
max-complexity = 10